1246-dev-create-minimum-working-time-series-example: Updated set_mapbox_apikey to refer to index.html now
