+++ Agents/TimeSeriesExample/dtvf_visualisation/img/icons/alpha-empty.png
