+++ Agents/TimeSeriesExample/dtvf_visualisation/img/icons/beta-empty.png
