+++ Agents/TimeSeriesExample/dtvf_visualisation/img/icons/gamma-empty.png
