+++ Agents/TimeSeriesExample/dtvf_visualisation/js/data_registry.js
@@ -8,78 +8,31 @@
 class DataRegistry {
 
     /**
-     * Overall meta data.
+     * Global metadata.
      */
-    _overallMeta;
+    _globalMeta = [];
 
     /**
-     * Metadata for fixed data.
+     * Metadata for visualisation data.
      */
-    _fixedMeta;
-
-    /**
-     * Metadata for additional data.
-     */
-    _additionalMeta = [];
+    _meta = [];
 
     /**
      * Returns the metadata for the entire visualisation.
      */
-    get overallMeta() {
-        return this._overallMeta;
-    }
-
-    /**
-     * Returns the metadata for the Fixed Data sets.
-     */
-    get fixedMeta() {
-        return this._fixedMeta;
+    get globalMeta() {
+        return this._globalMeta;
     }
 
     /**
      * Returns the metadata for the Additional Data sets.
      */
-    get additionalMeta() {
-        return this._additionalMeta;
+    get meta() {
+        return this._meta;
     }
 
     /**
-     * Returns the location of the Fixed Data directory.
-     */
-    getFixedDirectory() {
-        let rootDir = this._overallMeta["rootDirectory"];
-        return rootDir + this._overallMeta["fixedDirectory"];
-    }
-
-    /**
-     * Returns the location of the directory represented by the input
-     * Additional Data set groups.
-     * 
-     * @param {string[]} groups 
-     */
-    getAdditionalDirectory(groups) {
-        let rootDir = this._overallMeta["rootDirectory"];
-        let additionalDirectory = this._overallMeta["additionalDirectory"];
-
-        return rootDir + additionalDirectory + "/" + groups.join("/");
-    }
-
-    /**
-     * Returns the metadata for the Additional Data set represented by the
-     * input groups.
-     * 
-     * @param {string[]} groups 
-     * 
-     * @returns metadata
-     */
-    getAdditionalGroup(groups) {
-        var result = [];
-        this.#recurseAdditional(this._additionalMeta, 0, groups, result);
-        return result[0];
-    }
-
-    /**
-     * Automatically identify and load data if that meets the standard
+     * Automatically identify and read metadata if that meets the standard
      * format outlined for Digital Twin visualisations.
      * 
      * Note that this method is asynchronous, so a callback should be 
@@ -90,55 +43,81 @@ class DataRegistry {
      */
     loadMetaData(rootDir, callback) {
         rootDir = (rootDir.endsWith("/")) ? rootDir : rootDir + "/";
-        let overallMeta = rootDir + "overall-meta.json";
+        let topFile = rootDir + "meta.json";
 
         // Reset data objects
-        this._overallMeta = null;
-        this._fixedMeta = null;
-        this._additionalMeta = [];
+        this._globalMeta = [];
+        this._meta = [];
 
-        // Load the overall meta
+        // Load the top level meta
         var registry = this;
-        var promise = $.getJSON(overallMeta, function(json) {
-            json["rootDirectory"] =  rootDir;
-            registry._overallMeta = json;
+        var promise = $.getJSON(topFile, function(data) {
+            // Store global
+            registry._globalMeta = data["global"];
+            registry._globalMeta["rootDirectory"] = rootDir;
+
+            // Store top level of metadata tree
+            registry._meta = data["local"];
         }).promise();
 
-        // After that's loaded, read the fixed and additional meta
-        promise.then(
-            function() {
-                var promises = [];
+        // Drill down and read the other levels
+        promise.then(() => {
+            var promises = [];
+            var groups = registry._meta["groups"];
 
-                // Read the fixed meta
-                var fixedPromise = registry.#loadFixedMeta(registry, rootDir);
-                promises.push(fixedPromise);
-                
-                if(registry._overallMeta["additionalDirectory"]) {
-                    var selectionsContainer = document.getElementById("selectionsContainer");
-                    if(selectionsContainer != null) selectionsContainer.style.display = "block";
-
-                    // If additional directories are present, recursively read that meta
-                    var additionalDir = rootDir + registry._overallMeta["additionalDirectory"];
-                    var additionalPromise = registry.#loadAdditionalMeta(registry, additionalDir, registry._additionalMeta);
-                    promises.push(additionalPromise);
-                } else {
-                    // No additional data sets, hide the selection controls
-                    var selectionsContainer = document.getElementById("selectionsContainer");
-                    if(selectionsContainer != null) selectionsContainer.style.display = "none";
-                }
-
-                // Once both of those are finished, fire the callback
-                Promise.all([promises]).then(() => {
-                    console.log("INFO: Metadata for the Fixed and Additional Data sets has been loaded.");
-                    if(callback != null) callback();
-                });              
-            },
-
-            function(error) {
-                // Hard fail
-                console.log("ERROR: Could not load metadata!");
+            for(var i = 0; i < groups.length; i++) {
+                var groupDirectory = rootDir + groups[i]["directory"];
+                var groupPromise = registry.#recurseReadMeta(registry, groupDirectory, groups[i]);
+                promises.push(groupPromise);
             }
-        );
+            
+            // Once both of those are finished, fire the callback
+            Promise.all(promises).then(() => {
+                console.log("INFO: Metadata has been loaded.");
+                if(callback != null) callback();
+            });   
+        });
+    }
+
+    /**
+     * Recurses depth first and returns the first leaf group.
+     * 
+     * @returns 
+     */
+    getFirstGroup() {
+        var group = [];
+        this.#recurseFirstGroup(this._meta, group);
+        return group;
+    }
+
+    /**
+     * Recurse depth wise to find first leaf group
+     * @param {*} currentEntry 
+     * @param {*} group 
+     */
+    #recurseFirstGroup(currentEntry, group) {
+        if(currentEntry["groups"]) {
+            if(currentEntry["groups"]["label"] && currentEntry["groups"]["groups"]) {
+                group.push(currentEntry["groups"]["groups"][0]["directory"]);
+                this.#recurseFirstGroup(currentEntry["groups"]["groups"][0], group);
+            } else {
+                group.push(currentEntry["groups"][0]["directory"]);
+                this.#recurseFirstGroup(currentEntry["groups"][0], group);
+            }
+        }
+    }
+
+    /**
+     * Returns the metadata object for the input grouping (e.g. ["scenario-0", "time-0"]);
+     * 
+     * @param {String[]} group group definition
+     * 
+     * @returns {JSONObject} metadata object
+     */
+    getGroup(group) {
+        var result = [];
+        this.#findGroup(this._meta, 0, group, result);
+        return result[0];
     }
 
     /**
@@ -150,7 +129,7 @@ class DataRegistry {
      * @param {string[]} groups names of group
      * @param {JSONObject[]} result container for result
      */
-      #recurseAdditional(currentGroup, i, groups, result) {
+    #findGroup(currentGroup, i, groups, result) {
         if(result.length > 0) return;
 
         if(currentGroup["directory"] === groups[i]) {
@@ -159,73 +138,63 @@ class DataRegistry {
 
             } else if(currentGroup["groups"]["groups"]) {
                for(var j = 0; j < currentGroup["groups"]["groups"].length; j++) {
-                    this.#recurseAdditional(currentGroup["groups"]["groups"][j], i + 1, groups, result);
+                    this.#findGroup(currentGroup["groups"]["groups"][j], i + 1, groups, result);
                 }
             }
+
         } else if(currentGroup["groups"]) {
-           for(var j = 0; j < currentGroup["groups"].length; j++) {
-                this.#recurseAdditional(currentGroup["groups"][j], i, groups, result);
+            var subgroups = currentGroup["groups"];
+            for(var j = 0; j < subgroups.length; j++) {
+                this.#findGroup(subgroups[j], i, groups, result);
             }
+
         } else if(currentGroup.length > 0) {
             for(var j = 0; j < currentGroup.length; j++) {
-                this.#recurseAdditional(currentGroup[j], i, groups, result);
+                this.#findGroup(currentGroup[j], i, groups, result);
             }
          }
     }
 
     /**
-     * Loads the metadata that describes the fixed data set.
-     * 
-     * @param {DigitalTwinDataLoader} self self
-     * @param {string} rootDir root directory of data structures.
-     * 
-     * @returns promise
-     */
-    #loadFixedMeta(self, rootDir) {
-        let fixedDirectory = self._overallMeta["fixedDirectory"];
-        let fixedMetaFile = rootDir + fixedDirectory + "/meta.json";
-
-        return $.getJSON(fixedMetaFile, function(json) {
-            self._fixedMeta = json;
-        }).promise();
-    }
-
-
-    /**
      * Recursively finds and loads the meta describing the additional
      * data sets.
      * 
-     * @param {DigitalTwinDataLoader} self self
+     * @param {DataRegistry} registry self
      * @param {string} parentDir parent directory
      * @param {JSONObject} parentEntry parent meta entry
      * 
      * @returns promise
      */
-    #loadAdditionalMeta(self, parentDir, parentEntry) {
+    #recurseReadMeta(registry, parentDir, parentEntry) {
         let metaFile = parentDir + "/meta.json";
+
+        // Read this meta file
         var promise = $.getJSON(metaFile, function(json) {
             return json;
         });
 
+        // Drill down
         return promise.then(
             function(json) {
-               
                 if(json["dataSets"]) {
                     parentEntry["dataSets"] = json["dataSets"];    
+                    parentEntry["thisDirectory"] = parentDir;    
 
                 } else if(json["groups"]) {
+                    parentEntry["thisDirectory"] = parentDir;
+
                     if(parentEntry.length == 0) {
                         parentEntry.push(json);
                     } else {
                         parentEntry["groups"] = json;
                     }
 
+                    // Recurse into subgroups and read metadata
                     var promises = [];
                     for(var i = 0; i < json["groups"].length; i++) {
                         let groupEntry = json["groups"][i];
                         let groupDir = parentDir + "/" + groupEntry["directory"];
-                        
-                        promises.push(self.#loadAdditionalMeta(self, groupDir, groupEntry));
+                        promises.push(registry.#recurseReadMeta(registry, groupDir, groupEntry));
                     }
                     return Promise.all(promises);
                 }
