--- Agents/TimeSeriesExample/dtvf_visualisation/queried_data/layer-tree.json
@@ -1,7 +1,0 @@
-[
-    {
-        "layerName": "Consumers",
-        "defaultState": "visible",
-        "layerIDs": ["consumers"]
-    }
-]
