--- Agents/TimeSeriesExample/dtvf_visualisation/queried_data/overall-meta.json
@@ -1,13 +1,0 @@
-{
-    "name": "DTVF Example Visualisation",
-    "description": "This is an example visualisation of the retrieved KG data.",
-
-    "apiKey": "",
-    "defaultCenter": [0.119115, 52.205363],
-    "defaultZoom": 13,
-    "defaultBearing": -30,
-    "defaultPitch": 40,
-
-    "selectionsTitle": "Current Group:",
-    "fixedDirectory": "fixed"
-}
