+++ Agents/TimeSeriesExample/dtvf_visualisation_old/css/digital_twin_style.css
@@ -191,8 +191,7 @@ h1, h2, h3, h4, h5, h6 {
 }
 #titleContainer {
 	width: 100%;
-	min-height: 40px;
-	max-height: 80px;
+	height: 40px;
 	text-align: center;
 	margin-bottom: 20px;
 }
@@ -397,7 +396,7 @@ h1, h2, h3, h4, h5, h6 {
 	border-radius: 8px;
 	padding: 10px;
 	margin-top: 10px;
-	min-height: 100px;
+	min-height: 150px;
 }
 #selectionsContainer p {
 	margin-bottom: 10px;
@@ -412,21 +411,6 @@ h1, h2, h3, h4, h5, h6 {
 	height: 25px;
 }
 #selectContainer select {
-	width: 69%;
-	height: 25px;
-	font-size: 9pt;
-}
-
-#rootSelectContainer {
-	height: 25px;
-	margin-top: 5px;
-}
-#rootSelectContainer label {
-	display: inline-block;
-	width: 29%;
-	height: 25px;
-}
-#rootSelectContainer select {
 	width: 69%;
 	height: 25px;
 	font-size: 9pt;
@@ -514,16 +498,4 @@ h1, h2, h3, h4, h5, h6 {
 	border-top: none;
   }
   
-  #no-meta-container {
-	  background-color: rgb(240, 240, 240);
-	  color: rgb(100, 100, 100);
-	  text-align: center;
-	  height: 100%;
-	  width: 100%;
-	  display: table;
-  }
-
-  #no-meta-container p {
-	display: table-cell;
-	vertical-align: middle;
-  }
+
