+++ Agents/TimeSeriesExample/dtvf_visualisation_old/img/arrow.png
