+++ Agents/TimeSeriesExample/dtvf_visualisation_old/img/collapse.png
