+++ Agents/TimeSeriesExample/dtvf_visualisation_old/js/panel_handler.js
