+++ Agents/TimeSeriesExample/dtvf_visualisation_old/local_icons.js
@@ -4,13 +4,6 @@
 // (as they'll change for each implementation) rather than within
 // the framework.
 const EXAMPLE_ICONS = [
-    "img/icons/circle-black.png",
     "img/icons/warning.png",
-    "img/icons/error.png",
-    "img/icons/alpha.png",
-    "img/icons/alpha-empty.png",
-    "img/icons/beta.png",
-    "img/icons/beta-empty.png",
-    "img/icons/gamma.png",
-    "img/icons/gamma-empty.png"
+    "img/icons/error.png"
 ];
