+++ Agents/TimeSeriesExample/resources/ts_example.properties
@@ -1,13 +1,13 @@
 ###--- Properties for Python script ---###
 
 # Output directory for (Geo)JSON files containing retrieved data
-output.directory = ./dtvf_visualisation/queried_data/fixed
+output.directory = ./dtvf_visualisation/data
 
 
 ###--- Properties for Digital Twin Visualisation Framework ---###
 
 # Mapbox API key (to be copied to "overall-meta.json" file)
-mapbox.apiKey = 
+mapbox.apiKey = pk.eyJ1IjoibWg4MDciLCJhIjoiY2t2czdmcHRrMHlwMTJycXAxeDU2Zm5oNyJ9.A4v205lJXE5EE7cAa9v4Fw
 
 
 ###--- Properties for JAVA TimeSeriesClient ---###
